umem: extract MmapArea raw OS allocator into umem/mmap.rs (#986 Phase 0)
First slice of the umem.rs decomposition. Phase 0 establishes the
`umem/` directory layout and pulls the raw OS-interaction code
out of the 2,647-line monolith.

Mechanical changes:
- `afxdp/umem.rs` → `afxdp/umem/mod.rs` (directory module).
- New `afxdp/umem/mmap.rs` (~125 LOC): MmapArea struct, the
  HUGE_PAGE_SIZE constant, the inherent impl with new() /
  as_nonnull_slice() / is_hugepage_backed() / slice() /
  slice_mut() / slice_mut_unchecked(), and the Drop impl that
  munmap()s the region on shutdown.
- `umem/mod.rs`: declares `pub(in crate::afxdp) mod mmap;` and
  re-exports MmapArea so consumers continue to reach it via
  `crate::afxdp::umem::MmapArea` (pre-move call sites unchanged).
- Visibility widened from pub(super) to pub(in crate::afxdp) on
  the items consumers reach across the new module boundary.

After this Phase 0:
- umem/mod.rs: 2,527 LOC (still a smell — Phase 1 will extract
  ring management; Phase 2 frame recycling; Phase 3 telemetry).
- umem/mmap.rs: 133 LOC.

No body changes. cargo test --bins 865/0/2.
