+++ userspace-dp/src/afxdp/umem/mod.rs
@@ -1,5 +1,8 @@
 use super::*;
 
+pub(super) mod mmap;
+pub(in crate::afxdp) use mmap::MmapArea;
+
 pub(super) struct WorkerUmemInner {
     area: MmapArea,
     umem: Umem,
@@ -85,18 +88,8 @@ impl WorkerUmemPool {
     }
 }
 
-pub(super) struct MmapArea {
-    ptr: NonNull<u8>,
-    /// Original requested size (passed to XSK via as_nonnull_slice).
-    len: usize,
-    /// Actual mmap size (may be rounded up for hugepage alignment).
-    mapped_len: usize,
-    /// Whether the region is backed by explicit 2 MB hugepages.
-    hugepage: bool,
-}
 
 /// 2 MB hugepage size.
-const HUGE_PAGE_SIZE: usize = 2 * 1024 * 1024;
 
 /// Hard capacity of the per-binding redirect inbox
 /// (`BindingLiveState::pending_tx`). Sized to cover the highest expected
@@ -402,117 +395,7 @@ impl OwnerProfilePeerWrites {
     }
 }
 
-impl MmapArea {
-    pub(super) fn new(len: usize) -> io::Result<Self> {
-        // Round up to 2 MB boundary for hugepage eligibility.
-        let aligned_len = (len + HUGE_PAGE_SIZE - 1) & !(HUGE_PAGE_SIZE - 1);
 
-        // Attempt 1: explicit 2 MB hugepages (requires system reservation).
-        let ptr = unsafe {
-            libc::mmap(
-                std::ptr::null_mut(),
-                aligned_len,
-                libc::PROT_READ | libc::PROT_WRITE,
-                libc::MAP_PRIVATE
-                    | libc::MAP_ANONYMOUS
-                    | libc::MAP_HUGETLB
-                    | libc::MAP_POPULATE
-                    | (21 << libc::MAP_HUGE_SHIFT), // MAP_HUGE_2MB
-                -1,
-                0,
-            )
-        };
-        if ptr != libc::MAP_FAILED {
-            let ptr = NonNull::new(ptr.cast::<u8>())
-                .ok_or_else(|| io::Error::other("null mmap pointer"))?;
-            eprintln!(
-                "xpf-ha: umem alloc {} bytes ({} MB, 2MB hugepages)",
-                aligned_len,
-                aligned_len / (1024 * 1024)
-            );
-            return Ok(Self {
-                ptr,
-                len,
-                mapped_len: aligned_len,
-                hugepage: true,
-            });
-        }
-
-        // Attempt 2: standard pages with MAP_POPULATE + THP advisory hint.
-        let ptr = unsafe {
-            libc::mmap(
-                std::ptr::null_mut(),
-                aligned_len,
-                libc::PROT_READ | libc::PROT_WRITE,
-                libc::MAP_PRIVATE | libc::MAP_ANONYMOUS | libc::MAP_POPULATE,
-                -1,
-                0,
-            )
-        };
-        if ptr == libc::MAP_FAILED {
-            return Err(io::Error::last_os_error());
-        }
-        // Request transparent hugepage backing (advisory, cannot fail).
-        unsafe {
-            libc::madvise(ptr, aligned_len, libc::MADV_HUGEPAGE);
-        }
-        let ptr =
-            NonNull::new(ptr.cast::<u8>()).ok_or_else(|| io::Error::other("null mmap pointer"))?;
-        eprintln!(
-            "xpf-ha: umem alloc {} bytes ({} MB, standard pages + THP hint)",
-            aligned_len,
-            aligned_len / (1024 * 1024)
-        );
-        Ok(Self {
-            ptr,
-            len,
-            mapped_len: aligned_len,
-            hugepage: false,
-        })
-    }
-
-    /// Returns the original requested length (for XSK registration).
-    pub(super) fn as_nonnull_slice(&self) -> NonNull<[u8]> {
-        NonNull::slice_from_raw_parts(self.ptr, self.len)
-    }
-
-    /// Whether this region is backed by explicit 2 MB hugepages.
-    #[cfg_attr(not(test), allow(dead_code))]
-    pub(super) fn is_hugepage_backed(&self) -> bool {
-        self.hugepage
-    }
-
-    pub(super) fn slice(&self, offset: usize, len: usize) -> Option<&[u8]> {
-        let end = offset.checked_add(len)?;
-        if end > self.len {
-            return None;
-        }
-        Some(unsafe { std::slice::from_raw_parts(self.ptr.as_ptr().add(offset), len) })
-    }
-
-    #[cfg_attr(not(test), allow(dead_code))]
-    pub(super) fn slice_mut(&mut self, offset: usize, len: usize) -> Option<&mut [u8]> {
-        unsafe { self.slice_mut_unchecked(offset, len) }
-    }
-
-    pub(super) unsafe fn slice_mut_unchecked(
-        &self,
-        offset: usize,
-        len: usize,
-    ) -> Option<&mut [u8]> {
-        let end = offset.checked_add(len)?;
-        if end > self.len {
-            return None;
-        }
-        Some(unsafe { std::slice::from_raw_parts_mut(self.ptr.as_ptr().add(offset), len) })
-    }
-}
-
-impl Drop for MmapArea {
-    fn drop(&mut self) {
-        let _ = unsafe { libc::munmap(self.ptr.as_ptr().cast::<c_void>(), self.mapped_len) };
-    }
-}
 
 #[cfg(test)]
 mod tests {
